+++ pySIMsalabim/experiments/CV.py
@@ -1,4 +1,6 @@
-"""Perform capacitance simulations"""
+"""Perform capacitance simulations
+Version: 1 (16-10-2024) ## Internal version number for this script. Refer to the changelog to review changes ## 
+"""
 ######### Package Imports #########################################################################
 
 import os, sys