+++ pySIMsalabim/experiments/JV_steady_state.py
@@ -1,4 +1,6 @@
-"""Perform steady-state JV simulations"""
+"""Perform steady-state JV simulations
+## Version: 1 (16-10-2024) ## Internal version number for this script. Refer to the changelog to review changes ## 
+"""
 
 ######### Package Imports #########################################################################